refactor: components/map becomes components/globe
The directory named a screen. The screen is gone, and every module still in it
is consumed by the globe: the terrain sources, the area drawing, the imagery
dates, the camera navigation, the overlay stack, the chrome and the backdrop.
`GlobeSurface` already lived one directory over, importing seven of its own
neighbours across a boundary that described nothing.

This repository's own guidelines argue the point about the Go package that was
called `backend`: a name for a position in a diagram rather than for a subject.
`map` had become the same thing.

Two comments went stale with it and are corrected: `mapTools` still described
its table as what the navigation column lists as children, and StudioScreen
pointed a reader at `components/map/MapSurface.tsx`, deleted two commits ago.
