+++ frontend/src/components/globe/MapChrome.tsx
