+++ frontend/src/components/globe/SpaceBackdrop.tsx
