+++ frontend/src/components/globe/terrain.ts
