+++ frontend/src/components/globe/useAreaDrawing.ts
